Separated two adapters for CMS and EWH into separate files
