+++ yql/essentials/core/histogram/eq_width_histogram.h
@@ -219,7 +219,7 @@ public:
     }
 
     // Seriailizes to a binary representation
-    std::pair<std::unique_ptr<char>, ui64> Serialize() const;
+    TString Serialize() const;
 
     void Aggregate(const TEqWidthHistogram& other);
